Refactor find tool window into common base class
This will allow the creation of custom navigation tool windows allowing
display and navigation to operators anywhere in the workflow.
